+++ src/main/java/com/bluecine/Blue_Cinema/repository/ChairRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChairRepository extends JpaRepository<Chair,Long>{
+public interface ChairRepository extends JpaRepository<Chair,Integer>{
     
 }
